minui: Always enable TW_USE_MINUI_CUSTOM_FONTS
TWRP wants to use this flag to become compatible with both LineageOS and
OmniROM. But in fact, the flag will not be defined correctly because
`minui.h` is included directly by Android.mk of healthd. This causes a
compile failure of healthd that `gr_sys_font()` and friends are not
declared.

Considering we're hosting our own clone of TWRP and have no need to be
compatible with other projects, we can always enable this flag.

Change-Id: I3a7cf2ff0108be2d4e4eab6e0ae907c249a994b6